1. Not sure if you wrote this yet?: Added functionality to query if a vehicle has been assigned to user. Once vehicle has been assigned and active is set to true on trip. We begin monitoring for vehicle location.
2. Vehicle Arriving Fragment: Added new Fragment which displays the registration number, rating and estimated time
till arrival for the vehicle assigned to the client. It also displays the vehicles location with a marker.
The controller for this fragment also monitors the vehicles location in respect to the client. Once they're within
10 meters of each other the vehicle is considered arrived.

3. Vehicle Arrived Fragment: Added a new Fragment that is displayed when the vehicle has arrived.
It displays the reg number of the vehicle and tells the user to hit continue once inside the vehicle.

4. Fully integrated functionality for monitoring the users location in respect to the destination.
Once the user and destination are within 10m of each other it considers the trip finished and begins the
next section of the trip process. Once the trip is considered finished it pulls the current balance of
the user from the database and deducts the cost of the trip from their balance and then pushes the updated
value back up.

5. Trip Finished Fragment: Added new Fragment that is displayed when the trip is over.
It notifys the client that the trip is concluded and that the cost has been deducted from their
balance. It then presents the user with two buttons finish. Which when pressed clears the polyline
and marker from the map and resets all the stored order details as the transaction is complete.
And a second option called Rate Us which if clicked starts a new fragment.

6. Rate Us Fragment: Added new Fragment that is displayed when the user clicks Rate Us on the previous
fragment. They are presented with a thumb up and a thumb down. We pull down the positive and negative line
counters for the vehicle they travelled in and based on the thumb pressed we increase either the positive or
negative line counter. Then it terminates the trip and calls the reset state function to clear all details.